Initial Setup of Debug Override Functionality
Setup a basic initial implementation of the override behavior with a custom drawer to allow for custom selection of the object that will be used as the debug handler. A simple example scene is included that prints out a number of Simpsons quotes with a Debug Logger with Blacklist functionality that is setup to omit 4 of the messages.